+++ CHANGELOG.md
@@ -33,6 +33,9 @@
 * Add type column to users settings permissions sets. Refs UIU-2167.
 * Add modal for 'Cash drawer reconciliation report'. Refs UIU-1959.
 * Provide `key` when rendering list elements. Refs UIU-2168.
+* Create 'Cash drawer reconciliation report' in PDF format. Refs UIU-2084.
+* Create 'Cash drawer reconciliation report' in CSV format. Refs UIU-1961.
+* Migrate to `renew-by-barcode`. Refs UIU-2096.
 * Move Tenant/Bursar exports to Users/Transfer criteria. Refs UIU-2098.
 
 ## [6.0.0](https://github.com/folio-org/ui-users/tree/v6.0.0) (2021-03-18)